% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_to_shp.R
\name{adjustment_to_shp}
\alias{adjustment_to_shp}
\title{Adjustment to shape}
\usage{
adjustment_to_shp(
  base_folder = "/home/dataop/script/EMIS/GMAPTRAFFIC/traffic_adjustment_module_v3/",
 
    adms_points_path = "/disk/v092.user/tlh/colleagues/jimmy_chan/adms_to_raster/data/lnglat_no_receptors/lnglat_no_receptors_hk_80.shp",
  datetime,
  format
)
}
\arguments{
\item{base_folder}{(character) Basefolder of tam results. Defaults to /home/dataop/script/EMIS/GMAPTRAFFIC/traffic_adjustment_module_v3/}

\item{adms_points_path}{(character) Path to the shapefile of ADMS points.}

\item{datetime}{(POSIXct) Datetime argument}

\item{format}{(character) Either "sf" or "terra", depending which package to use}
}
\value{
An object of class \code{sfc_multipoint} or \code{vect}
}
\description{
Reads in adjustment (from .csv) and generates a shapefile
}
\keyword{adjustment}
\keyword{shapefile}
