% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adms.R
\name{read_adms}
\alias{read_adms}
\title{Reads in ADMS result to dataframe}
\usage{
read_adms(
  datetime,
  adms_version = "hkv20b",
  pollutants = c("PM10", "PM2.5", "NO2", "O3", "SO2"),
  base_folder = "/home/dataop/data/nmodel/adms"
)
}
\arguments{
\item{datetime}{(POSIXct) Datetime argument.}

\item{adms_version}{(character) e.g. "hkv20b"}

\item{pollutants}{(character vector) e.g. c("PM10", "O3"), can be all variables of the .nc file}

\item{base_folder}{(character) Basefolder of ADMS results. Default is "/home/dataop/data/nmodel/adms"}

\item{adms_points_path}{(character) Path to the shapefile of ADMS points.}
}
\value{
An object of class \code{data.frame}
}
\description{
Reads in ADMS result (from .nc file) and outputs a dataframe
}
\keyword{adms}
