% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adms.R
\name{read_adms_poins}
\alias{read_adms_poins}
\title{Reads in the shapefile that has the information where ADMS output points are located}
\usage{
read_adms_poins(
 
    adms_points_path = "/disk/v092.user/tlh/colleagues/jimmy_chan/adms_to_raster/data/lnglat_no_receptors/lnglat_no_receptors_hk_80.shp"
)
}
\arguments{
\item{adms_points_path}{(character) Path to the shapefile of ADMS points.}
}
\value{
Returns nothing, but sets 'adms_shp_dummy' as a \code{sf} object in the global environment
}
\description{
Reads in ADMS result (from .nc file) and outputs a dataframe
}
