% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adms.R
\name{read_adms_raster}
\alias{read_adms_raster}
\title{Reads in ADMS result to shapefile}
\usage{
read_adms_raster(
  datetime,
  resolution,
  adms_version = "hkv20b",
  pollutants = c("PM10", "PM2.5", "NO2", "O3", "SO2"),
  base_folder = "/home/dataop/data/nmodel/adms",
 
    adms_points_path = "/disk/v092.user/tlh/colleagues/jimmy_chan/adms_to_raster/data/lnglat_no_receptors/lnglat_no_receptors_hk_80.shp",
  cores = length(pollutants)
)
}
\arguments{
\item{datetime}{(POSIXct) Datetime argument.}

\item{resolution}{(numeric) Resolution of raster (in m).}

\item{adms_version}{(character) e.g. "hkv20b"}

\item{pollutants}{(character vector) e.g. c("PM10", "O3"), can be all variables of the .nc file}

\item{base_folder}{(character) Basefolder of ADMS results. Default is "/home/dataop/data/nmodel/adms"}

\item{adms_points_path}{(character) Path to the shapefile of ADMS points.}

\item{cores}{(numeric) Number of cores for rasterization. Default is one core per pollutant.}
}
\value{
An object of class \code{sf}
}
\description{
Reads in ADMS result (from .nc file) and outputs a dataframe
}
\keyword{adms}
