% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upl_emissions_to_df.R
\name{sources_in_upl}
\alias{sources_in_upl}
\title{Lists all sources in an .upl file}
\usage{
sources_in_upl(UPL_PATH)
}
\arguments{
\item{UPL_PATH}{(character) Full path to the UPL file}
}
\value{
An object of class \code{character}
}
\description{
Lists all sources in an .upl file and gives back a character vector
}
\keyword{adms}
