% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upl_emissions_to_df.R
\name{upl_emissions_to_df}
\alias{upl_emissions_to_df}
\title{Reads the emissions of an upl file as a dataframe}
\usage{
upl_emissions_to_df(UPL_PATH, ROUTE_IDs, cores = 10)
}
\arguments{
\item{UPL_PATH}{(character) Full path to the UPL file}

\item{ROUTE_IDs}{(character vector) Which route ids should be read in?}

\item{adms_points_path}{(character) Path to the shapefile of ADMS points.}
}
\value{
An object of class \code{data.frame}
}
\description{
Reads the emissions of an upl file and outputs a dataframe
}
\keyword{adms}
